CEDS-1860 Extract Messages from ViewSpec into MessagesStub
