+++ test/views/ViewSpec.scala
@@ -16,56 +16,25 @@
 
 package views
 
-import base.Injector
+import base.MessagesStub
 import controllers.CSRFSupport
 import controllers.exchanges.{AuthenticatedRequest, JourneyRequest, Operator}
 import models.cache.Answers
 import org.jsoup.Jsoup
 import org.jsoup.nodes.{Document, Element}
 import org.scalatest.{MustMatchers, WordSpec}
-import play.api.i18n.{Lang, Messages, MessagesApi}
 import play.api.mvc.{AnyContent, Request}
 import play.api.test.FakeRequest
 import play.twirl.api.Html
 import testdata.CommonTestData.providerId
 
-class ViewSpec extends WordSpec with MustMatchers with ViewTemplates with ViewMatchers with Injector with CSRFSupport {
+class ViewSpec extends WordSpec with MustMatchers with ViewTemplates with ViewMatchers with MessagesStub with CSRFSupport {
 
   implicit protected def htmlBodyOf(html: Html): Document = Jsoup.parse(html.toString())
-
-  private val messagesApi: MessagesApi = ViewSpec.realMessagesApi
-
-  protected implicit def messages(implicit request: Request[_]): Messages =
-    new AllMessageKeysAreMandatoryMessages(messagesApi.preferred(request))
-
-  protected def messages(key: String, args: Any*)(implicit request: Request[_]): String = messages(request)(key, args: _*)
 
   protected implicit val fakeRequest: Request[AnyContent] = FakeRequest().withCSRFToken
 
   protected def journeyRequest(answers: Answers) = JourneyRequest(answers, AuthenticatedRequest(Operator(providerId), fakeRequest))
-
-  /*
-    Fails the test if a view is configured with a message key that doesnt exist in the messages file
-   */
-  private class AllMessageKeysAreMandatoryMessages(msg: Messages) extends Messages {
-    override def messages: Messages = msg.messages
-
-    override def lang: Lang = msg.lang
-
-    override def apply(key: String, args: Any*): String =
-      if (msg.isDefinedAt(key))
-        msg.apply(key, args: _*)
-      else throw new AssertionError(s"Message Key is not configured for {$key}")
-
-    override def apply(keys: Seq[String], args: Any*): String =
-      if (keys.exists(key => !msg.isDefinedAt(key)))
-        msg.apply(keys, args)
-      else throw new AssertionError(s"Message Key is not configured for {$keys}")
-
-    override def translate(key: String, args: Seq[Any]): Option[String] = msg.translate(key, args)
-
-    override def isDefinedAt(key: String): Boolean = msg.isDefinedAt(key)
-  }
 
   /*
     Implicit Utility class for retrieving common elements which are on the vast majority of pages
@@ -84,8 +53,4 @@ class ViewSpec extends WordSpec with MustMatchers with ViewTemplates with ViewMa
     def getForm: Option[Element] = Option(document.getElementsByTag("form")).filter(!_.isEmpty).map(_.first())
   }
 
-}
-
-object ViewSpec extends Injector {
-  val realMessagesApi = instanceOf[MessagesApi]
 }
